test: Disable IPv6 in kube-proxy tests for externalTrafficPolicy=Local
Kubernetes only seems to have full support for IPv6 with 1.20. With
previous versions, the entries for externalTrafficPolicy=Local services
are missing in the kube-proxy iptables rules.

This commits therefore disables IPv6 in kube-proxy tests with
externalTrafficPolicy=Local services if running with K8s < 1.20.

The helper functions (testNodePort and testExternalTrafficPolicyLocal)
have been updated to probe for IPv6 support in the cilium-agent via
cilium status.

Signed-off-by: Paul Chaignon <paul@cilium.io>