cilium: auto-detect network interfaces
We need to auto-detect network interfaces in the EKS environment that by
default uses three interfaces eth0, eth1, and eth2. For an initial effort
just pattern match the EKS case and attach to any network interfaces with
eth* in the name.

In general we could attempt to inspect the routing table, but that also
feels fragile depending on routing rules. This patch resolves the EKS
case and leaves other cases for future work.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>