bpf: add classification key & mask struct
Add a v4 hash table map as well as key (mask) and value structures. The
idea is to emulate a poor man's wildcard flow lookup via hash table map
until we have a proper wildcard map support in the kernel. The rule_id
will be passed into the perf RB so that the user space agent can pick
it up and correlate the matched rule.

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>